% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glag.r
\name{glag}
\alias{glag}
\title{Generalized Lag Function}
\usage{
glag(x, k)
}
\arguments{
\item{x}{An array.}

\item{k}{A shift number. Can be potitive or negative integer.}
}
\description{
Compute a lagged version of a series or matrix, shifting the series by a given number.
The shift number can be either positive or negative. Appropriate number of NA's are added
either at the begging or end of the returned series so that it is the same length as the
original series.
}
\examples{
x <- 1:10
glag(x, 3)
glag(MonthName(x), -3)
glag(matrix(1:24,nrow=6, ncol=4), 2)
}

