% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.r
\name{\%notin\%}
\alias{\%notin\%}
\title{Not-In}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{vector or NULL: the values to be matched.}

\item{table}{vector or NULL: the values to be matched against.}
}
\description{
The negation of %in%, which returns a logical vector indicating if there is a non-match or not for its left operand.
}
\examples{
1:10 \%notin\% c(1,3,5,9)
}

