% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/between.r
\name{isBetween}
\alias{isBetween}
\title{Identify values between two values}
\usage{
isBetween(x, start, end)
}
\arguments{
\item{x}{A numeric value or an array of values or dates.}

\item{start}{A numeric value or a date.}

\item{end}{A numeric value or a date.}
}
\description{
Identify values between two values
}
\examples{
isBetween(3:10, start=1, end=5)
isBetween(as.Date("2012-1-1"), startDate, endDate)
}

