% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DailyAndMonthly.r
\name{ConvertDailyToMonthly}
\alias{ConvertDailyToMonthly}
\alias{ConvertMonthlyToDaily}
\title{Convert between Daily and Monthly Data Frames}
\usage{
ConvertDailyToMonthly(df)

ConvertMonthlyToDaily(df, type = "step")
}
\arguments{
\item{df}{A data frame to be converted. The first column needs to be a date column.}

\item{type}{Smoothing type for monthly-to-daily conversion. c("step", "smooth"). Default set at "step."}
}
\description{
ConvertDailyToMonthly takes a daily table and convert it to a monthly table with values equal to the mean.
Conversely, ConvertMonthlyToDaily takes a monthly table and convert it to a daily table, either as a step function or
pricewise-defined 2nd-degree polynomials.
}
\section{Functions}{
\itemize{
\item \code{ConvertMonthlyToDaily}: Take a monthly data frame and convert it to daily data frame.
}}
\examples{
niter <- 2
monthly.df <- data.frame("Date"=rep(seq(as.Date("2010-1-1"), by="month", length.out = 12),niter)
, "Iter" = rep(1:niter,each=12), "Y"=rnorm(12*niter))
monthly.df <- cast(monthly.df, Date ~ Iter, value="Y")
plot0(monthly.df$Date, monthly.df$'1', ylim=c(-3,3))
matlines(monthly.df$Date, monthly.df[,-1], lwd=2, lty=1)

daily.df <- ConvertMonthlyToDaily(monthly.df)
matlines(daily.df$Date, daily.df[,-1], lty=2)

daily.df <- ConvertMonthlyToDaily(monthly.df, type="smooth")
matlines(daily.df$Date, daily.df[,-1])

verify.df <- ConvertDailyToMonthly(daily.df)
matlines(verify.df$Date, verify.df[,-1], lwd=10, lty=1, col=gray.f(0.3))
}

